+++ sources/EngineWorks/DBAL/Sqlite/DBAL.php
@@ -4,6 +4,7 @@ use EngineWorks\DBAL\CommonTypes;
 use EngineWorks\DBAL\DBAL as AbstractDBAL;
 use EngineWorks\DBAL\Traits\MethodSqlIsNull;
 use EngineWorks\DBAL\Traits\MethodSqlLike;
+use EngineWorks\DBAL\Traits\MethodSqlLimit;
 use EngineWorks\DBAL\Traits\MethodSqlQuote;
 use SQLite3;
 
@@ -12,6 +13,8 @@ class DBAL extends AbstractDBAL
     use MethodSqlQuote;
     use MethodSqlLike;
     use MethodSqlIsNull;
+    use MethodSqlLimit;
+
     /**
      * Contains the connection resource for SQLite3
      * @var SQLite3
@@ -156,15 +159,6 @@ class DBAL extends AbstractDBAL
     public function sqlIfNull($fieldName, $nullValue)
     {
         return 'IFNULL(' . $fieldName . ', ' . $nullValue . ')';
-    }
-
-    public function sqlLimit($query, $requestedPage, $recordsPerPage = 20)
-    {
-        $rpp = max(1, $recordsPerPage);
-        $query = rtrim($query, "; \t\n\r\0\x0B")
-            . ' LIMIT ' . $this->sqlQuote($rpp * (max(1, $requestedPage) - 1), CommonTypes::TINT)
-            . ', ' . $this->sqlQuote($rpp, CommonTypes::TINT) . ';';
-        return $query;
     }
 
     public function sqlRandomFunc()